Create Basic Structure of Page
Created html and css files to display the title at the top of the page,
"Rock, Paper, Scissors", a middle section with "YOU", r-p-s options,
"CPU" with a blank box beneath it, and a comparison box in between
(you is on left, cpu is on right). Also has a bottom section where the
winner will be stated after the game is played.
These are all placeholder boes for now and they are outlined as such.