Comment out body
I realized after reading through the assignment that the game is
going to be played entirely through the consol. No css required.
So I commented out the body on the html file so nothing is displayed
on the page, but the code is there for future use after the assignment
is done as asked and can be improved upon in the future by giving it
a proper GUI.