Change Random Number Generator for Computer Pick
Changed the way the computer player randomly selects rock, paper, or
scissors. The old method was getting a random number 0-.999 and
choosing rock if < .333, paper if < .666 and scissors if greater.
Now the random number is from 0 - 2 by using Math.floor(Math.random()*3)
which will pick number 0, 1, or 2 which correlates to rock, paper, or
scissors for the computer player.