Create UI Elements for RPS Game
Created three buttons for rock, paper, and scissors for the user to
press to make their selection. The game logic has also been changed
so that it makes a check with what the computer randomly chose only
when the user presses one of the buttons. When either the user or the
computer gets to 5 points, the game ends and the buttons are removed.
The running score is displayed at all times.